its_multi_agent/backend/app/requirements.txt
fastapi
uvicorn
python-dotenv
pymysql
DBUtils
openai
httpx
openai-agents
pystun3

# OpenTelemetry 链路追踪
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
